requirements.txt
python-telegram-bot==21.6
anthropic==0.34.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
